Improve repodata.txt github actions
Allows all non-hidden (starting with '.') folders and all non-hidden
files to be catalogued with a repodata.txt file.

This enables remote access for raw as well as database files, and allows
github actions to track this whole repo without needing to add lines
to the .yaml file whenever a new database or raw file (or whatever else)
is added.